+++ CG_Engine/Properties.h
@@ -7,8 +7,8 @@
 namespace GL_Engine {
 	namespace Properties {
 		struct GLFWproperties {
-			uint16_t		width;
-			uint16_t		height;
+			uint32_t		width;
+			uint32_t		height;
 			const char*		title;
 			GLFWmonitor*	monitor;
 			GLFWwindow*		window;
